requirements.txt
streamlit>=1.34
requests>=2.31
pandas>=2.2
numpy>=1.26
google-generativeai>=0.4.0     
opencv-python-headless>=4.10     
Pillow>=10.3                    
plotly>=5.22
